Close gh-3
The decrypted and uncompressed stream was copied to a memory stream which was then being copied to the file stream. However, since the memory stream was not rewound to its beginning we ended up writing no data. On the other hand we increased the amount of data written by the length of the memory stream so the reported extracted size was correct and of course non-zero.